Address Greptile review round 6: Decimal consistency, dead code, API key docs
- Remove float() cast from settlement_value in kalshi settlements (P1)
- Remove dead skip_count variable and unused CursorResult import (P2)
- Document API key query-param constraint for PW/VC/OWM clients (P2)
- Change bracket_low/bracket_high from Float to Numeric(10,4) with migration (P2)
- Update parse_bracket() and all tests to use Decimal for bracket values

+++ shared/db/models.py
@@ -134,8 +134,8 @@ class KalshiMarket(Base):
         Enum(MarketType, name="market_type_enum", create_constraint=True),
         nullable=False,
     )
-    bracket_low: Mapped[float | None] = mapped_column(Float, nullable=True)
-    bracket_high: Mapped[float | None] = mapped_column(Float, nullable=True)
+    bracket_low: Mapped[Decimal | None] = mapped_column(Numeric(10, 4), nullable=True)
+    bracket_high: Mapped[Decimal | None] = mapped_column(Numeric(10, 4), nullable=True)
     is_edge_bracket: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     status: Mapped[MarketStatus] = mapped_column(
         Enum(MarketStatus, name="market_status_enum", create_constraint=True),
